pyproject.toml
[project]
name = "iaai-p"
version = "0.1.0"
description = ""
authors = [
    {name = "Tomasz Szłapa",email = "tomasz.szlapa@gmail.com"}
]
requires-python = ">=3.11"
dependencies = [
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)"
]
